refactor: remove duplicated impl
