+++ crates/rspack_plugin_lightning_css_minimizer/Cargo.toml
@@ -8,17 +8,17 @@ version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-lightningcss = { workspace = true }
 rayon        = { workspace = true }
 regex        = { workspace = true }
 thread_local = { workspace = true }
 tracing      = { workspace = true }
 
-rspack_core  = { workspace = true }
-rspack_error = { workspace = true }
-rspack_hash  = { workspace = true }
-rspack_hook  = { workspace = true }
-rspack_util  = { workspace = true }
+rspack_core                = { workspace = true }
+rspack_error               = { workspace = true }
+rspack_hash                = { workspace = true }
+rspack_hook                = { workspace = true }
+rspack_loader_lightningcss = { workspace = true }
+rspack_util                = { workspace = true }
 
 [package.metadata.cargo-shear]
 ignored = ["tracing"]
